services/workflow-engine/pyproject.toml
[project]
name = "workflow-engine"
version = "0.1.0"
description = "Workflow engine service for the NGEN platform"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi>=0.111,<1.0",
    "uvicorn>=0.30,<1.0",
    "ngen-framework-core",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "httpx>=0.27,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/workflow_engine"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
